CosyVoice/requirements_simple.txt
--extra-index-url https://download.pytorch.org/whl/cu121
conformer==0.3.2
diffusers==0.29.0
gdown==5.1.0
gradio==5.4.0
hydra-core==1.3.2
HyperPyYAML==1.2.2
inflect==7.3.1
librosa==0.10.2
lightning==2.2.4
matplotlib==3.7.5
modelscope==1.20.0
networkx==3.1
omegaconf==2.3.0
onnx==1.16.0
onnxruntime==1.18.0
openai-whisper==20231117
protobuf==4.25
pydantic==2.7.0
rich==13.7.1
soundfile==0.12.1
torch==2.3.1
torchaudio==2.3.1
transformers==4.51.3
wetext==0.0.4
wget==3.2
